src/neural/attention/pattern/quantum.py
"""Quantum pattern dynamics implementation."""

from typing import Optional, Tuple, List
import torch
import numpy as np
from dataclasses import dataclass


@dataclass
class QuantumState:
    """Quantum state representation for patterns."""
    
    amplitude: torch.Tensor  # Complex amplitude tensor
    phase: torch.Tensor  # Phase tensor
    
    @property
    def state_vector(self) -> torch.Tensor:
        """Get full quantum state vector."""
        return self.amplitude * torch.exp(1j * self.phase)
    
    def evolve(self, hamiltonian: torch.Tensor, dt: float) -> 'QuantumState':
        """Evolve quantum state using Hamiltonian.
        
        Args:
            hamiltonian: Hamiltonian operator
            dt: Time step
            
        Returns:
            Evolved quantum state
        """
        # Convert to complex tensor
        state = self.state_vector
        
        # Time evolution operator
        U = torch.matrix_exp(-1j * hamiltonian * dt)
        
        # Evolve state
        evolved = torch.matmul(U, state)
        
        # Extract new amplitude and phase
        new_amplitude = torch.abs(evolved)
        new_phase = torch.angle(evolved)
        
        return QuantumState(new_amplitude, new_phase)


class QuantumGeometricTensor:
    """Quantum geometric tensor implementation."""
    
    def __init__(self, dim: int):
        """Initialize quantum geometric tensor.
        
        Args:
            dim: Dimension of parameter space
        """
        self.dim = dim
        
    def compute_tensor(self, state: QuantumState) -> torch.Tensor:
        """Compute quantum geometric tensor.
        
        Args:
            state: Quantum state
            
        Returns:
            Quantum geometric tensor
        """
        # Get state vector
        psi = state.state_vector
        
        # Initialize tensor
        Q = torch.zeros((self.dim, self.dim), dtype=torch.complex64)
        
        # Compute tensor components
        for i in range(self.dim):
            for j in range(self.dim):
                # Compute derivatives
                dpsi_i = self._parameter_derivative(psi, i)
                dpsi_j = self._parameter_derivative(psi, j)
                
                # Compute tensor element
                Q[i,j] = torch.sum(torch.conj(dpsi_i) * dpsi_j)
                
        return Q
    
    def decompose(self, Q: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
        """Decompose quantum geometric tensor into metric and Berry curvature.
        
        Args:
            Q: Quantum geometric tensor
            
        Returns:
            Tuple of (metric tensor, Berry curvature)
        """
        # Metric is real part (symmetric)
        g = torch.real(Q)
        
        # Berry curvature is imaginary part (antisymmetric)
        B = torch.imag(Q)
        
        return g, B
    
    def _parameter_derivative(self, state: torch.Tensor, param_idx: int) -> torch.Tensor:
        """Compute parameter derivative of state.
        
        Args:
            state: Quantum state vector
            param_idx: Parameter index
            
        Returns:
            Parameter derivative
        """
        # Use finite differences for now
        # TODO: Implement analytic derivatives
        eps = 1e-6
        param_shift = torch.zeros(self.dim)
        param_shift[param_idx] = eps
        
        state_plus = self._apply_param_shift(state, param_shift)
        state_minus = self._apply_param_shift(state, -param_shift)
        
        return (state_plus - state_minus) / (2 * eps)
    
    def _apply_param_shift(self, state: torch.Tensor, shift: torch.Tensor) -> torch.Tensor:
        """Apply parameter shift to state.
        
        Args:
            state: Quantum state vector
            shift: Parameter shift vector
            
        Returns:
            Shifted state
        """
        # For now just apply phase shift
        # TODO: Implement proper parameter transformations
        phase = torch.sum(shift)
        return state * torch.exp(1j * phase) 